# openapi.yaml
openapi: 3.0.0
info:
  title: WeatherOps API
  description: >
    Event-driven weather intelligence platform for real-time monitoring,
    alert management, and automated weather-triggered actions.

    ## Features
    - User authentication with JWT tokens
    - Weather location monitoring
    - Custom weather alert rules
    - Real-time alert management
    - Multi-channel notifications
    - User preferences and settings

    ## Base URL
    Production: `https://api.weatherops.local/api/v1`
    Development: `http://localhost:8000/api/v1`
  version: 1.1.0
  contact:
    name: WeatherOps Support
    email: support@weatherops.local
  license:
    name: Apache 2.0
  x-logo:
    url: https://weatherops.local/logo.png
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://api.weatherops.local/api/v1
    description: Production server
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
  schemas:
...
    AlertResponse:
      type: object
      required:
        - id
        - location_id
        - rule_id
        - user_id
        - metric
        - actual_value
        - threshold
        - operator
        - severity
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Alert unique identifier
          example: 550e8400-e29b-41d4-a716-446655440000
        location_id:
          type: string
          format: uuid
          description: Associated location ID
          example: 660e8400-e29b-41d4-a716-446655440000
        rule_id:
          type: string
          format: uuid
          description: Associated rule ID
          example: 770e8400-e29b-41d4-a716-446655440000
        user_id:
          type: string
          format: uuid
          description: User ID
          example: 880e8400-e29b-41d4-a716-446655440000
        metric:
          type: string
          enum:
            - temperature
            - rainfall
            - wind_speed
            - humidity
          description: Weather metric that triggered alert
          example: temperature
        actual_value:
          type: number
          format: double
          description: Current measured value
          example: 38.5
        threshold:
          type: number
          format: double
          description: Configured threshold value
          example: 35.0
        operator:
          type: string
          enum:
            - ">"
            - "<"
            - ">="
            - "<="
            - "=="
          description: Comparison operator used
          example: ">"
        severity:
          type: string
          enum:
            - LOW
            - MEDIUM
            - HIGH
          description: Alert severity level
          example: HIGH
        status:
          type: string
          enum:
            - active
            - resolved
          description: Current alert status
          example: active
        weather_snapshot:
          type: string
          nullable: true
          description: JSON snapshot of weather data at alert time
          example: null
        created_at:
          type: string
          format: date-time
          description: Alert creation timestamp
          example: 2024-06-05T10:30:00Z
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
          example: 2024-06-05T10:30:00Z
        resolved_at:
          type: string
          format: date-time
          nullable: true
          description: When alert was resolved
          example: null
      description: Weather alert information response
    AlertResolve:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - active
            - resolved
          description: New alert status
          example: resolved
      description: Alert status update request

    # User Preference Schemas
...
    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message
          example: Invalid credentials
      description: Error response
    ValidationErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  oneOf:
                    - type: string
                    - type: integer
              msg:
                type: string
              type:
                type: string
          description: Validation error details
      description: Validation error response
...
paths:
...
  /alerts:
    get:
      tags:
        - Alerts
      summary: List Alerts
      description: Get weather alerts with optional filtering and pagination
      parameters:
        - name: location_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Filter by location UUID
        - name: severity
          in: query
          required: false
          schema:
            type: string
            enum:
              - LOW
              - MEDIUM
              - HIGH
          description: Filter by severity level
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum:
              - active
              - resolved
          description: Filter by status
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
          description: Maximum number of results
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of results to skip
      responses:
        '200':
          description: Alerts retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AlertResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: get_alerts
  /alerts/location/{location_id}:
    get:
      tags:
        - Alerts
      summary: Get Location Alerts
      description: Get active alerts for a specific location
      parameters:
        - name: location_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Location ID
        - name: severity
          in: query
          required: false
          schema:
            type: string
            enum:
              - LOW
              - MEDIUM
              - HIGH
          description: Filter by severity
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 50
          description: Maximum number of results
      responses:
        '200':
          description: Location alerts retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AlertResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Location not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: get_location_alerts
  /alerts/{alert_id}:
    get:
      tags:
        - Alerts
      summary: Get Alert
      description: Get details for a specific alert
      parameters:
        - name: alert_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Alert ID
      responses:
        '200':
          description: Alert retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Alert not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: get_alert

    patch:
      tags:
        - Alerts
      summary: Update Alert
      description: Update alert status (resolve or reactivate)
      parameters:
        - name: alert_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Alert ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertResolve'
      responses:
        '200':
          description: Alert updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Alert not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
      operationId: update_alert

    delete:
      tags:
        - Alerts
      summary: Delete Alert
      description: Delete an alert record
      parameters:
        - name: alert_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Alert ID
      responses:
        '204':
          description: Alert deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Alert not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: delete_alert
...